+++ changelog.md
@@ -63,5 +63,6 @@
 
 ## Next
 
+* Tests clear require cache before each test
 * Run jshint on test code
 * Restructure tests